+++ src/containers/ConfirmationContainer/ConfirmationContainer.vue
@@ -49,6 +49,8 @@
       ref="successModal"
       :message="successMessage"
       :link-message="linkMessage"
+      :link-to="linkTo"
+      :etherscan-link="etherscanLink"
     />
     <error-modal
       ref="errorModal"
@@ -128,6 +130,8 @@ export default {
       signedMessage: '',
       successMessage: 'Success',
       linkMessage: 'OK',
+      linkTo: '/',
+      etherscanLink: null,
       dismissed: true,
       signedArray: [],
       txBatch: null,
@@ -156,10 +160,13 @@ export default {
     });
   },
   created() {
-    this.$eventHub.$on('showSuccessModal', (message, linkMessage) => {
-      if (!message) message = null;
-      this.showSuccessModal(message, linkMessage);
-    });
+    this.$eventHub.$on(
+      'showSuccessModal',
+      (message, linkMessage, etherscanLink) => {
+        if (!message) message = null;
+        this.showSuccessModal(message, linkMessage, etherscanLink);
+      }
+    );
 
     this.$eventHub.$on('showErrorModal', (message, linkMessage) => {
       if (!message) message = null;
@@ -329,10 +336,11 @@ export default {
       window.scrollTo(0, 0);
       this.$refs.signConfirmModal.$refs.signConfirmation.show();
     },
-    showSuccessModal(message, linkMessage) {
+    showSuccessModal(message, linkMessage, etherscanLink) {
       this.reset();
       if (message !== null) this.successMessage = message;
       if (linkMessage !== null) this.linkMessage = linkMessage;
+      if (etherscanLink !== null) this.etherscanLink = etherscanLink;
       this.$refs.successModal.$refs.success.show();
     },
     showErrorModal(message, linkMessage) {
@@ -438,8 +446,16 @@ export default {
       this.dismissed = false;
       this.responseFunction(this.signedTxObject);
       this.$refs.confirmModal.$refs.confirmation.hide();
+
       if (this.raw.generateOnly) return;
-      this.showSuccessModal('Transaction sent!', 'Okay');
+      this.showSuccessModal(
+        'Transaction sent!',
+        'Okay',
+        this.network.type.blockExplorerTX.replace(
+          '[[txHash]]',
+          this.signedTxObject.tx.hash
+        )
+      );
     },
     reset() {
       this.responseFunction = null;